Changed Bird, BirdAction, BirdActionInterface, BirdInstance, Game, Player
Bird: changed the form of image loading to make it lazy loading to same RAM. Also added deckCount decreaser and updated performAction to include gameContext.
BirdAction: added gameContext to the methods and functions and finished ALLDRAW1BIRD function.
BirdActionInterface: updated it to add gameContext
BirdInstance: updated it to add gameContext
Game: Added playerlist arrayLIst and return method for it
Player: Changed birdHand and corresponding methods and constructor to keep as Bird. Plan to instantiate bird to BirdInstance when placed on board.